requirements.txt
# MQTT Client
paho-mqtt==1.6.1

# HTTP Client
requests==2.31.0

# WebSocket Client
websocket-client==1.6.4

# JSON handling and utilities
urllib3==2.0.7

# Optional: For better logging and configuration
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-mock==3.12.0
